# src/config/swagger.yaml
openapi: 3.0.0
info:
  title: todoList
  description: todoList app API
  version: 0.1.9
servers:
  - url: https://todo-backend-app.up.railway.app/api
    description: Production server
  - url: http://localhost:3000/api
    description: Local server
paths:
...
  /deleteTodo/{id}:
    delete:
      tags:
        - todo
      summary: Deletes a todo item
      operationId: deleteTodo
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Todo deleted successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
...